+++ Assets/Scripts/NPC/NPCQuest.cs
@@ -31,7 +31,7 @@ public class NPCQuest : NPCCharacter {
             QuestCamera.SetActive (true);
         }
     }
-    public override void OnHit (int val) {
+    public override void OnHit (float val) {
         throw new NPCShootedException ("YOU HIT: NPC_QUEST");
     }
 }
